refactor(frontend): rename custom-instance.ts to api-client.ts
custom-instance.ts was an Orval-internal naming convention that poorly
described the file's domain purpose. Rename to api-client.ts, which
clearly conveys that this is the HTTP client consumed by generated hooks.

Update orval.config.ts mutator path, regenerate all generated imports,
and fix the sole manual reference in LoginForm.tsx.
